+++ src/frontend/app/pages/about.tsx
@@ -1,0 +1,7 @@
+export default function AboutPage() {
+  return (
+      <div className="bg-base-100 border border-base-300">
+        <div className="grid place-content-center h-80">Expression calculator</div>
+      </div>
+  )
+}